# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ClienteResponse:
    properties:
      cpf:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
      telefone:
        type: string
      updated_at:
        type: string
    type: object
  dto.CountResponse:
    properties:
      count:
        type: integer
    type: object
...
  dto.CreateItemPedidoRequest:
    properties:
      produto_id:
        type: integer
      quantidade:
        type: integer
    required:
    - produto_id
    - quantidade
    type: object
  dto.CreatePedidoRequest:
    properties:
      cliente_id:
        type: integer
      itens:
        items:
          $ref: '#/definitions/dto.CreateItemPedidoRequest'
        minItems: 1
        type: array
      status:
        enum:
        - pendente
        - pago
        - enviado
        - entregue
        - cancelado
        type: string
    required:
    - cliente_id
    - itens
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  dto.ItemPedidoResponse:
    properties:
      id:
        type: integer
      preco_unitario:
        type: number
      produto:
        $ref: '#/definitions/dto.ProdutoResponse'
      produto_id:
        type: integer
      quantidade:
        type: integer
      subtotal:
        type: number
    type: object
  dto.PedidoResponse:
    properties:
      cliente:
        $ref: '#/definitions/dto.ClienteResponse'
      cliente_id:
        type: integer
      created_at:
        type: string
      data_pedido:
        type: string
      id:
        type: integer
      itens:
        items:
          $ref: '#/definitions/dto.ItemPedidoResponse'
        type: array
      status:
        type: string
      updated_at:
        type: string
      valor_total:
        type: number
    type: object
  dto.ProdutoResponse:
    properties:
      ativo:
        type: boolean
      categoria:
        type: string
      created_at:
        type: string
      descricao:
        type: string
      estoque:
        type: integer
      id:
        type: integer
      nome:
        type: string
      preco:
        type: number
      sku:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdatePedidoRequest:
    properties:
      status:
        enum:
        - pendente
        - pago
        - enviado
        - entregue
        - cancelado
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@exemplo.com
    name: API Support
  description: REST API for managing clientes (customers) with CRUD operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cliente API
  version: "1.0"
paths:
...
  /pedidos:
    get:
      description: Retrieve all pedidos from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PedidoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all pedidos
      tags:
      - pedidos
    post:
      consumes:
      - application/json
      description: Create a new pedido with the provided information
      parameters:
      - description: Pedido data
        in: body
        name: pedido
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePedidoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PedidoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new pedido
      tags:
      - pedidos
  /pedidos/{id}:
    delete:
      description: Delete a pedido by ID
      parameters:
      - description: Pedido ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete pedido
      tags:
      - pedidos
    get:
      description: Retrieve a specific pedido by ID
      parameters:
      - description: Pedido ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PedidoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get pedido by ID
      tags:
      - pedidos
    put:
      consumes:
      - application/json
      description: Update the status of an existing pedido
      parameters:
      - description: Pedido ID
        in: path
        name: id
        required: true
        type: integer
      - description: Pedido status update
        in: body
        name: pedido
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePedidoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PedidoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update pedido status
      tags:
      - pedidos
  /pedidos/cliente/{cliente_id}:
    get:
      description: Retrieve all pedidos for a specific cliente
      parameters:
      - description: Cliente ID
        in: path
        name: cliente_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PedidoResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get pedidos by cliente ID
      tags:
      - pedidos
  /pedidos/count:
    get:
      description: Get the total number of pedidos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Count pedidos
      tags:
      - pedidos
  /pedidos/status/{status}:
    get:
      description: Retrieve pedidos by status (pendente, pago, enviado, entregue,
        cancelado)
      parameters:
      - description: Pedido status
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PedidoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get pedidos by status
      tags:
      - pedidos
...
swagger: "2.0"